Problem 771 Solution
Description: There are two lists, one with Jewels, and one with stones. The jewels are unique, the stones may carry Jewels. Determine how many jewels exist in the stone list. e.g. J = [a,A] s=[a,a,a,B,B,B] --> 3 stones are Jewels
Language: JS
Runtime: 64ms (Could be better...)
Comment: I'll try and improve the algorithm later, but as it stands, the approach is clear and readable enough.